[vim] Fix SignColumn colours.
Change SignColumn to use bg_base02 so it blends in with the fold and
number columns.

Also add a series of other highlighting colours for the sign column to
allow Syntastic to hook into them.  This is not an ideal solution, since
it relies on two otherwise unrelated plugins (Solarized and Sytanstic)
sharing knowledge of non-standard highlighting targets.